+++ src/components/Avatar/types/common.ts
@@ -1,7 +1,5 @@
-import type {DOMProps} from '../../types';
-
 export type AvatarSize = '3xs' | '2xs' | 'xs' | 's' | 'm' | 'l' | 'xl';
 
-export interface AvatarCommonProps extends Pick<DOMProps, 'className'> {
+export interface AvatarCommonProps {
     size: AvatarSize;
 }