+++ src/components/FilePreview/__stories__/FilePreview.stories.tsx
@@ -3,10 +3,13 @@ import {action} from '@storybook/addon-actions';
 import type {Meta, StoryFn} from '@storybook/react';
 
 import {Icon} from '../../Icon';
-import {Flex} from '../../layout';
+import {Text} from '../../Text';
+import {Flex} from '../../layout/Flex/Flex';
 import type {FilePreviewProps} from '../FilePreview';
 import {FilePreview} from '../FilePreview';
 import {FILE_TYPES} from '../types';
+
+import {CompactMenuWithActionsShowcase} from './CompactFilePreviewWithActionsShowcase/CompactFilePreviewWithActionsShowcase';
 
 export default {
     title: 'Components/Data Display/FilePreview',
@@ -116,17 +119,15 @@ const noClickableTemplateActions = [
     },
 ];
 
-const NoClickableTemplate: StoryFn<Omit<FilePreviewProps, 'actions'>> = (args) => {
+const NoClickableTemplate: StoryFn<Omit<FilePreviewProps, 'actions'>> = () => {
     return (
         <Flex gap={4}>
-            <FilePreview {...args} file={{name: 'No clickable', type: 'text/docs'} as File} />
+            <FilePreview file={{name: 'No clickable', type: 'text/docs'} as File} />
             <FilePreview
-                {...args}
                 file={{name: 'No clickable with actions', type: 'text/docs'} as File}
                 actions={noClickableTemplateActions}
             />
             <FilePreview
-                {...args}
                 file={{name: 'Clickable with actions', type: 'text/docs'} as File}
                 onClick={() => action('onClick')}
                 actions={noClickableTemplateActions}
@@ -136,6 +137,43 @@ const NoClickableTemplate: StoryFn<Omit<FilePreviewProps, 'actions'>> = (args) =
 };
 
 export const NoClickable = NoClickableTemplate.bind({});
+
+const SelectedTemplate: StoryFn<Omit<FilePreviewProps, 'actions'>> = () => {
+    return (
+        <div
+            style={{
+                display: 'grid',
+                justifyItems: 'center',
+                alignItems: 'center',
+                gridTemplateColumns: 'repeat(3, 1fr)',
+                gridGap: 8,
+                width: 400,
+            }}
+        >
+            <Text>File</Text>
+            <FilePreview selected file={{name: 'file', type: 'text/docs'} as File} />
+            <FilePreview
+                selected
+                view="compact"
+                file={{name: 'Size s file', type: 'text/docs'} as File}
+            />
+            <Text>Image</Text>
+            <FilePreview
+                selected
+                file={{name: 'image', type: 'image/png'} as File}
+                imageSrc="https://storage.yandexcloud.net/uikit-storybook-assets/changelog-dialog-picture-2.png"
+            />
+            <FilePreview
+                selected
+                view="compact"
+                file={{name: 'image', type: 'image/png'} as File}
+                imageSrc="https://storage.yandexcloud.net/uikit-storybook-assets/changelog-dialog-picture-2.png"
+            />
+        </div>
+    );
+};
+
+export const Selected = SelectedTemplate.bind({});
 
 const withoutActionTooltipTemplateActions = [
     {
@@ -148,11 +186,10 @@ const withoutActionTooltipTemplateActions = [
     },
 ];
 
-const WithoutActionTooltipTemplate: StoryFn<Omit<FilePreviewProps, 'actions'>> = (args) => {
+const WithoutActionTooltipTemplate: StoryFn<Omit<FilePreviewProps, 'actions'>> = () => {
     return (
         <Flex gap={4}>
             <FilePreview
-                {...args}
                 file={{name: 'Clicable without tooltip', type: 'text/docs'} as File}
                 onClick={() => action('onClick')}
                 actions={withoutActionTooltipTemplateActions}
@@ -162,3 +199,59 @@ const WithoutActionTooltipTemplate: StoryFn<Omit<FilePreviewProps, 'actions'>> =
 };
 
 export const WithoutActionTooltip = WithoutActionTooltipTemplate.bind({});
+
+const ViewTemplate: StoryFn<Omit<FilePreviewProps, 'actions'>> = () => {
+    return (
+        <div
+            style={{
+                display: 'grid',
+                justifyItems: 'center',
+                alignItems: 'center',
+                gridTemplateColumns: 'repeat(3, 1fr)',
+                gridGap: 8,
+                width: 400,
+            }}
+        >
+            <Text>Default</Text>
+            <FilePreview file={{name: 'file', type: 'text/docs'} as File} />
+            <FilePreview
+                file={{name: 'image', type: 'image/png'} as File}
+                imageSrc="https://storage.yandexcloud.net/uikit-storybook-assets/changelog-dialog-picture-2.png"
+            />
+            <Text>Compact</Text>
+            <FilePreview view="compact" file={{name: 'Size s file', type: 'text/docs'} as File} />
+            <FilePreview
+                view="compact"
+                file={{name: 'image', type: 'image/png'} as File}
+                imageSrc="https://storage.yandexcloud.net/uikit-storybook-assets/changelog-dialog-picture-2.png"
+            />
+        </div>
+    );
+};
+
+export const View = ViewTemplate.bind({});
+
+const compactMenuActions = [
+    {
+        iconStart: <Link width={14} height={14} />,
+        text: 'open on drive',
+        action: () => action('onLink'),
+    },
+    {
+        iconStart: <Icon data={Xmark} width={14} height={14} />,
+        text: 'delete a file',
+        action: () => action('onClose'),
+    },
+];
+
+const CustomMenuInDesktopCompactModeTemplate: StoryFn<Omit<FilePreviewProps, 'actions'>> = () => {
+    return (
+        <CompactMenuWithActionsShowcase
+            file={{name: 'image with actions', type: 'image/png'} as File}
+            imageSrc="https://storage.yandexcloud.net/uikit-storybook-assets/changelog-dialog-picture-2.png"
+            actions={compactMenuActions}
+        />
+    );
+};
+
+export const CustomMenuInDesktopCompactMode = CustomMenuInDesktopCompactModeTemplate.bind({});