+++ src/components/Radio/__stories__/RadioShowcase.tsx
@@ -8,13 +8,18 @@ export function RadioShowcase() {
             <ShowcaseItem title="Size">
                 <div style={{marginBlock: '1em'}}>
                     <Radio size="m" content="size m" value="value 1" />
-                    <span style={{margin: '8px'}} />
+                    <span style={{margin: '16px'}} />
                     <Radio size="m" content="size m" value="value 2" />
                 </div>
                 <div style={{marginBlock: '1em'}}>
                     <Radio size="l" content="size l" value="value 1" />
-                    <span style={{margin: '8px'}} />
+                    <span style={{margin: '16px'}} />
                     <Radio size="l" content="size l" value="value 2" />
+                </div>
+                <div style={{marginBlock: '1em'}}>
+                    <Radio size="xl" content="size xl" value="value 1" />
+                    <span style={{margin: '8px'}} />
+                    <Radio size="xl" content="size xl" value="value 2" />
                 </div>
             </ShowcaseItem>
             <ShowcaseItem title="Disabled">