+++ src/components/SegmentedRadioGroup/__stories__/SegmentedRadioGroup.stories.tsx
@@ -1,8 +1,10 @@
+import {CircleInfoFill, TriangleExclamationFill} from '@gravity-ui/icons';
 import {action} from '@storybook/addon-actions';
 import type {Meta, StoryObj} from '@storybook/react';
 
 import {Showcase} from '../../../demo/Showcase';
 import {ShowcaseItem} from '../../../demo/ShowcaseItem';
+import {Icon} from '../../Icon';
 import {SegmentedRadioGroup} from '../SegmentedRadioGroup';
 
 import {SegmentedRadioGroupShowcase} from './SegmentedRadioGroupShowcase';
@@ -19,6 +21,25 @@ export const Default: Story = {
         children: [
             <SegmentedRadioGroup.Option key="Value 1" value="Value 1" content="Value 1" />,
             <SegmentedRadioGroup.Option key="Value 2" value="Value 2" content="Value 2" />,
+            <SegmentedRadioGroup.Option key="Value 3" value="Value 3" content="Value 3" />,
+        ],
+        defaultValue: 'Value 1',
+        onUpdate: action('onUpdate'),
+        onFocus: action('onFocus'),
+        onBlur: action('onBlur'),
+    },
+};
+
+export const Icons: Story = {
+    args: {
+        children: [
+            <SegmentedRadioGroup.Option key="Value 1" value="Value 1" title="Warning">
+                <Icon data={TriangleExclamationFill} />
+                <span>Warning</span>
+            </SegmentedRadioGroup.Option>,
+            <SegmentedRadioGroup.Option key="Value 2" value="Value 2" title="Info">
+                <Icon data={CircleInfoFill} />
+            </SegmentedRadioGroup.Option>,
             <SegmentedRadioGroup.Option key="Value 3" value="Value 3" content="Value 3" />,
         ],
         defaultValue: 'Value 1',